requirements.txt
datasets~=2.16.1
evaluate~=0.4.1
scikit-learn~=1.3.2
pandas~=2.1.4
numpy~=1.26.2
matplotlib~=3.8.2
seaborn~=0.13.0
protobuf~=5.29.4
python-dotenv~=1.0.0
torch~=2.1.2
transformers~=4.36.2
sentence-transformers~=2.2.2
setfit~=1.0.3
spacy~=3.7.2
unidecode~=1.3.7
psycopg2~=2.9.9
gensim~=4.3.2
requests~=2.31.0
snorkel~=0.9.9